refactor: initView
